requirements.txt
keras==3.2.0
tensorflow==2.17.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-hub
keras_cv